changed hooks to utility functions
